+++ torchmetrics/functional/regression/__init__.py
@@ -23,4 +23,5 @@ from torchmetrics.functional.regression.mean_squared_error import mean_squared_e
 from torchmetrics.functional.regression.mean_squared_log_error import mean_squared_log_error  # noqa: F401
 from torchmetrics.functional.regression.pearson import pearson_corrcoef  # noqa: F401
 from torchmetrics.functional.regression.r2 import r2_score  # noqa: F401
+from torchmetrics.functional.regression.r2score import r2score  # noqa: F401
 from torchmetrics.functional.regression.spearman import spearman_corrcoef  # noqa: F401
